requirements.txt
Pillow
numpy
torch
transformers
pytorch_lightning
gradio
